[move-only] Move header info for borrow to destructure transform into its own header and out of MoveOnlyObjectChecker.h
