+++ lib/SILOptimizer/Mandatory/MoveOnlyBorrowToDestructureTransformTester.cpp
@@ -20,6 +20,7 @@
 
 #define DEBUG_TYPE "sil-move-only-checker"
 
+#include "MoveOnlyBorrowToDestructure.h"
 #include "MoveOnlyDiagnostics.h"
 #include "MoveOnlyObjectChecker.h"
